src/SimGCF/requirements.txt
tf-keras>=2.12.0
tqdm>=4.65.0
sentence-transformers>=2.2.2
torch_geometric>=2.3.0
wandb>=0.12.0
numpy==1.24.4
rust
polars
#pip install torch-scatter -f https://data.pyg.org/whl/torch-2.7.0+cpu.html